+++ src/components/atoms/HyperLink.js
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 
 import { Link as GatsbyLink } from 'gatsby';
 
-import { hyperLinkColor } from 'constants/palette'
+import { hyperLinkColor } from 'constants/palette';
 
 export default function HyperLink({ to, children }) {
   const LinkStyle = styled.div`
