+++ src/pages/404.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import SEO from 'templates/SEO';
 import BaseLayout from 'templates/BaseLayout';
-import HyperLink from 'components/molecules/HyperLink';
+import HyperLink from 'components/atoms/HyperLink';
 import { Text, H1Text } from 'components/atoms/Text';
 import Accent from 'components/atoms/Accent';
